requirements.txt
# requirements.txt
numpy~=2.2.0
Pillow~=11.0.0
scikit-learn~=1.6.0
matplotlib~=3.10.0
opencv-python~=4.10.0.84
pip~=24.3.1
setuptools~=75.6.0
scipy~=1.14.1
threadpoolctl~=3.5.0
joblib~=1.4.2
six~=1.17.0
python-dateutil~=2.9.0.post0
contourpy~=1.3.1
fonttools~=4.55.3
packaging~=24.2
pyparsing~=3.2.0
cycler~=0.12.1
kiwisolver~=1.4.7